refactor(vcs): move rulesets into the github provider package
Rulesets are GitHub-specific, so they do not belong alongside the
provider-agnostic abstractions.
